更正缩进

+++ src/com/mcml/space/doevent/AntiSpam.java
@@ -11,13 +11,13 @@ import org.bukkit.event.player.AsyncPlayerChatEvent;
 import com.mcml.space.core.VLagger;
 
 public class AntiSpam
-        implements Listener {
+implements Listener {
 
     private final static HashMap<String, Long> CheckList = new HashMap<String, Long>();
 
     private static boolean CheckFast(String bs) {
         if (CheckList.containsKey(bs)) {
-            return (((Long) CheckList.get(bs)).longValue() + VLagger.AntiSpamPeriodPeriod * 1000 > System.currentTimeMillis());
+            return (CheckList.get(bs).longValue() + VLagger.AntiSpamPeriodPeriod * 1000 > System.currentTimeMillis());
         }
         return false;
     }
@@ -34,11 +34,11 @@ public class AntiSpam
                 event.setCancelled(true);
                 p.sendMessage(VLagger.PluginPrefix + VLagger.AntiSpamPeriodWarnMessage);
             }else{
-            	CheckList.put(pn, System.currentTimeMillis());
+                CheckList.put(pn, System.currentTimeMillis());
             }
         }
     }
-    
+
     @EventHandler
     public void AntiDirty(AsyncPlayerChatEvent event) {
         if (VLagger.AntiSpamenable == true) {
@@ -49,11 +49,11 @@ public class AntiSpam
             List<String> strings = VLagger.AntiSpamDirtyList;
             int ss = strings.size();
             for(int i = 0;i < ss;i++){
-            	String string = strings.get(i);
-            	if(event.getMessage().contains(string)){
-            		event.setCancelled(true);
-            		p.sendMessage(VLagger.PluginPrefix + VLagger.AntiSpamDirtyWarnMessage);
-            	}
+                String string = strings.get(i);
+                if(event.getMessage().contains(string)){
+                    event.setCancelled(true);
+                    p.sendMessage(VLagger.PluginPrefix + VLagger.AntiSpamDirtyWarnMessage);
+                }
             }
         }
     }
